Use criterion instead of nightly feature
