+++ mast/src/lib.rs
@@ -1,5 +1,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
-#![feature(test)]
 
 #[cfg(not(feature = "std"))]
 extern crate alloc;
